subindo novas atualizações

+++ libft/ft_strlcpy.c
@@ -6,11 +6,13 @@
 /*   By: gigardin <gigardin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/20 20:05:05 by gigardin          #+#    #+#             */
-/*   Updated: 2023/07/22 20:26:06 by gigardin         ###   ########.fr       */
+/*   Updated: 2023/07/27 18:18:22 by gigardin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlcpy(char *dst, char *src, size_t size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dst, char *src, size_t size)
 {
 	size_t	index;
